cast {{{sizeof()}}} to {{{unsigned long}}} in a couple of places to eliminate warnings under Cygwin; refs #90

+++ miscSupport/logp.cc
@@ -118,10 +118,10 @@ main()
 	 tmp3.unlog(),
 	 tmp4.unlog());
 
-  printf("sizeof(logp<float,float>) = %lu\n",sizeof(logp<float,float>));
-  printf("sizeof(logp<double,float>) = %lu\n",sizeof(logp<double,float>));
-  printf("sizeof(logp<float,double>) = %lu\n",sizeof(logp<float,double>));
-  printf("sizeof(logp<double,double>) = %lu\n",sizeof(logp<double,double>));
+  printf("sizeof(logp<float,float>) = %lu\n",(unsigned long)sizeof(logp<float,float>));
+  printf("sizeof(logp<double,float>) = %lu\n",(unsigned long)sizeof(logp<double,float>));
+  printf("sizeof(logp<float,double>) = %lu\n",(unsigned long)sizeof(logp<float,double>));
+  printf("sizeof(logp<double,double>) = %lu\n",(unsigned long)sizeof(logp<double,double>));
 
 
   foo(tmp,tmp);
